+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 3.3.4 (August 29, 2025)
 -----------------------
+- Fix #393: Update module resources path
 - Fix #400: Fixed missing `not saved` warning
 
 3.3.3 (August 23, 2025)